Module Wtr

Types

type 'a router

'a t represents a Trie based router.

and 'c route

'c route is a uri and its handler. 'c represents the value returned by the handler.

and ('a, 'b) uri

('a, 'b) uri represents a route URI - both the path and query, e.g. /home/about/, /home/contact, /home/contact?name=a&no=123 etc. It is created via %wtr ppx

and method' = [

| `GET

| `HEAD

| `POST

| `PUT

| `DELETE

| `CONNECT

| `OPTIONS

| `TRACE

| `Method of string

]

method' represents HTTP request methods. It can be used as part of a uri in %wtr ppx.

and 'a decoder

Represents a uri component decoder, such as :int, :float, :bool etc.

Specifying a URI

Specifying a URI in a %wtr ppx follows the following syntax:

wtr syntax = http methods separated by comma ';' http uri

uri syntax = http uri

http uri = HTTP request path syntax

http uri = HTTP request path syntax ? request query

A URI in a %wtr ppx is syntactically and sematically a HTTP URI with the addition of decoders and some some useful additions listed below:

The default HTTP method if none is specified is GET method.

Full splat ** - Full spat operator matches any/all path following a full splat. For example /home/** matches the following uri paths, /home/about/, home/contact, /home/product etc. Full splat must be the last component of an uri. It is an error to specify other uri path component after full splat operator. Additionally wtr decodes the remaining matched url in the route handler. For example,

let r =
  Wtr.t [{%wtr|get; /public/** |} (fun url -> Format.sprintf "%s" url)]
in
let s = Wtr.match' `GET "/public/css/style.css" in
s = Some "css/style.css"

Wildward * - A wildcard operator matches any text appearing on the path component position. For example, uri /home/*/page1 matches the following /home/23/page1, /home/true/page1, /home/234.4/page1 etc. The semantics of wildcard operator is the same as using :string decoder in a uri, i.e. it affects the route handler function signature.

Trailing slash / - A trailing slash ensures that Wtr will match a trailing / in a uri. For example, uri /home/about/ matches /home/about/ but not /home/about.

Decoders

Built-in Decoders

Wtr provides the following built in decoders that can be used as when specifying URI in {%wtr| |} ppx:

:int - decodes a int

:int32 - decodes a int32

:int64 - decodes a int64

:float - decodes a float or int

:bool - decodes a bool

:string - decodes a string

The built-in decoders can be used as follows:

{%wtr|get; /home/:int |}, {%wtr| /home/:bool |}

Custom Decoders

val decoder : name:string -> decode:(string -> 'a option) -> 'a decoder

decoder ~name ~decode creates a user defined custom uri decoder component. name is used during the pretty printing of uri.

Wtr supports creating custom, user defined decoders. The convention for user defined decoders is as follows:

A custom decoder should be defined in a module which has the following:

type called `t`

a value called t which is of type t Wtr.decoder.

Custom decoder example,

module Fruit = struct
  type t = Apple | Orange | Pineapple

  let t : t Wtr.decoder =
    Wtr.decoder ~name:"fruit" ~decode:(function
      | "apple" -> Some Apple
      | "orange" -> Some Orange
      | "pineapple" -> Some Pineapple
      | _ -> None )
end

The custom decoder thus defined can then can be used in %wtr ppx as follows:

{%wtr| get ; /fruit/:Fruit |} fruit_page

Route Handlers

Route handlers are functions that accepts the decoded data from URI. A HTTP method, a URI and a route handler makes a route. The use of decoders \- both built-in and custom - affect the function signature of a route handler. For e.g.

A uri spec /home/:int/:bool expects a route handler as fun (i:int) (b:bool) -> ....

A uri spec /home/:string expects a route handler as (fun (s:string) -> ...)

Router

val router : 'a route list list -> 'a router

route routes is a router made up of given routes. It is used with %wtr ppx.

Wtr.router [{%wtr| /public/**|}] 

val match' : method' -> string -> 'a router -> 'a option

match' method' request_target router is Some a if method' and request_target together matches one of the routes defined in router. Otherwise it is None.

HTTP Method

val method_equal : method' -> method' -> bool

val method' : string -> method'

Pretty Printers

val pp : Stdlib.Format.formatter -> 'a router -> unit

val pp_method : Stdlib.Format.formatter -> method' -> unit

val pp_route : Stdlib.Format.formatter -> 'b route -> unit