Module Wtr

A HTTP request routing library.

Wtr (Well Typed Router) is a HTTP request routing library for OCaml web applications.

Given a HTTP request_target and a HTTP method', wtr attempts to match the two properties to a pre-defined set of routes. If a match is found then the corresponding route handler function of the matched route is executed.

The well typed part in Wtr means that the route handler functions can capture and receive arguments which are typed in a variety of OCaml types.

wtr implements two Request Target DSLs to aid in specifying routes for the router:

Combinators based

PPX based - provided by opam package wtr-ppx

References

RFC 7230 - HTTP Request Target

RFC 7231 - HTTP Methods

Types

type 'a router

A router consists of one or many HTTP request routes which are used to match a given HTTP request target. The base algorithm used for matching is a radix trie.

'a represents the value returned after executing the corresponding route handler of a matched route.

and 'a route

route is a HTTP request route. A route encapsulates a HTTP method', a request_target and a route handler. A route handler is either of the following:

a value of type 'a

or a function which returns a value of type 'a.

and ('a, 'b) request_target

request_target is a HTTP request target value to be matched. It consists of either just a path value or a combination of path and query values.

Example request_target values:

/home/about/ - path only

/home/contact - path only

/home/contact?name=a&no=123 - path (/home/contact) and query (name=a&no=123). Path and query are delimited by ? character token if both are specified.

Consult Request Target DSL for creating values of this type.

See HTTP RFC 7230 - request target.

and ('a, 'b) path

path is a part of request_target. It consists of one or more path components. path components are tokens which are delimited by a / character token.

Example of path and path components:

/ has path a component /

/home/about has path components home, about

/home/contact/ has path components home, contact and /

Consult path combinators for creating values of this type.

and ('a, 'b) query

query is a part of request_target. It consists of one of more query components which are delimited by a & character token. A query component further consists of a pair of values called name and value. name and value tokens are delimited by a = character token. A query component is represented syntactically as (name,value).

Given a request_target /home/about?a=2&b=3, the query components are (a,2) and (b,3).

Consult query combinators for creating values of this type.

and method' = [

| `GET

| `HEAD

| `POST

| `PUT

| `DELETE

| `CONNECT

| `OPTIONS

| `TRACE

| `Method of string

]

method' is a HTTP request method. See HTTP RFC 7231 - HTTP Methods

and 'a arg

arg is a component which can convert a path component or a query component value token into an OCaml typed value represented by 'a. The successfully converted value is then fed to a route handler function as an argument.

Arg

val arg : string -> (string -> 'a option) -> 'a arg

arg name convert is arg with a name name and convert as the function which will convert/decode a string value to an OCaml value of type 'a.

name is used during the pretty-printing of request_target by pp_request_target.

convert v is Some a if convert can successfully convert v to a. Otherwise it is None.

The following defines an arg of type Fruit.t arg.

module Fruit = struct
  type t = Apple | Orange | Pineapple

  let t : t Wtr.arg =
    Wtr.arg "fruit" (function
      | "apple" -> Some Apple
      | "orange" -> Some Orange
      | "pineapple" -> Some Pineapple
      | _ -> None )
end

Request Target DSL

Request Target combinators implement a DSL(domain specific language) to specify request_target values and hence both path component and query component values.

Let's assume that we want to specify a HTTP route which matches a request target value as such:

match a string literal "hello" exactly

followed by a valid OCaml int value

and then finally followed by an OCaml string value

We can use path combinators to implement such a requirement:

let target1 = Wtr.(exact "hello" / int / string /. pend) 

The target1 above matches the following instances of HTTP request target:

/home/2/str1

/home/3/str2

/home/-10/str3

General Components

val (/) : (('a, 'b) path -> 'c) -> ('d -> ('a, 'b) path) -> 'd -> 'c

p1 / p2 is a closure that combines p1 and p2. p1 and p2 are closures which encapsulate path value.

val (/&) : (('a, 'b) query -> 'c) -> ('d -> ('a, 'b) query) -> 'd -> 'c

q1 /& q1 is a closure that combines q1 and q2. q1 and q2 are closures which encapsulate query value.

val (/?) : (('a, 'b) path -> 'c) -> ('d -> ('a, 'b) query) -> 'd -> 'c

p /? q is a closure which combines p and q. p is a closure which encapsulates path value and q is a closure which encapsulate query value.

val (/.) : (('d, 'e) path -> ('b, 'c) path) -> ('d, 'e) path -> ('b, 'c) request_target

p /. pe is a request_target value that consists of only path components. pe is a path value that match the last path component.

val (/?.) : (('b, 'b) query -> ('c, 'd) path) -> unit -> ('c, 'd) request_target

pq /?. () is a request_target value. pq is a closure which encapulates both path and query components.

let request_target1 =
  Wtr.(
    exact "hello"
    / bool
    /? qint "hello"
    /& qstring "hh"
    /& qbool "b"
    /?. ())

val exact : string -> ('a, 'b) path -> ('a, 'b) path

exact e p matches a path component to e exactly.

val qexact : (string * string) -> ('a, 'b) query -> ('a, 'b) query

qexact (field, e) matches a query component to e exactly. The query component token name value is field.

val to_request_target : ('a, 'b) path -> ('a, 'b) request_target

to_request_target p is request_target consisting of only path p.

val root : ('a, 'a) request_target

root is a request_target with / as the only component, i.e. it matches exactly the root HTTP request.

Arg Components

Path/Query arg components encapsulate arg value which are then fed to a route handler function as an argument.

Path

val int : ('a, 'b) path -> (int -> 'a, 'b) path

int matches valid OCaml int values.

val int32 : ('a, 'b) path -> (int32 -> 'a, 'b) path

int32 matches valid OCaml int32 values.

val int64 : ('a, 'b) path -> (int64 -> 'a, 'b) path

int64 matches valid OCaml int64 values.

val float : ('a, 'b) path -> (float -> 'a, 'b) path

float matches valid OCaml float values.

val bool : ('a, 'b) path -> (bool -> 'a, 'b) path

bool matches valid OCaml bool values.

val string : ('a, 'b) path -> (string -> 'a, 'b) path

string matches valid OCaml string values.

val parg : 'c arg -> ('a, 'b) path -> ('c -> 'a, 'b) path

parg d p matches a path component if arg d can successfully convert path component to a value of type 'c.

Query

val qint : string -> ('a, 'b) query -> (int -> 'a, 'b) query

qint field matches a valid OCaml int value. field is the name token of query component.

val qint32 : string -> ('a, 'b) query -> (int32 -> 'a, 'b) query

qint32 field matches a valid OCaml int32 value. field is the name token of query component.

val qint64 : string -> ('a, 'b) query -> (int64 -> 'a, 'b) query

qint64 field matches a valid OCaml int64 value. field is the name token of query component.

val qfloat : string -> ('a, 'b) query -> (float -> 'a, 'b) query

qfloat field matches a valid OCaml float value. field is the name token of query component.

val qbool : string -> ('a, 'b) query -> (bool -> 'a, 'b) query

qbool field matches a valid OCaml bool value. field is the name token of query component.

val qstring : string -> ('a, 'b) query -> (string -> 'a, 'b) query

qstring field matches a valid OCaml string value. field is the name token of query component.

val qarg : (string * 'c arg) -> ('a, 'b) query -> ('c -> 'a, 'b) query

qarg (field, d) matches a query component if arg d can successfully convert path component to a value of type 'c.

Matching Last Components

These combinators match the last(end) path component. They are used with (/.) function.

val pend : ('a, 'a) path

pend matches the end of path value.

val splat : (string -> 'a, 'a) path

splat matches and captures all of the remaining path and query components. The captured value is then fed to a route handler.

val slash : ('a, 'a) path

slash matches path component / first and then matches the end of the path value.

HTTP Method

val method_equal : method' -> method' -> bool

method_equal m1 m2 is true if m1 and m2 is the same value. Otherwise it is false.

Note if both m1 and m2 are `Method m then the string comparison is case insensitive.

Wtr.method_equal `GET `GET = true;;
Wtr.method_equal `POST `GET = false;;
Wtr.method_equal (`Method "meth") (`Method "METH") = true

val method' : string -> method'

method' m is method' where string value m is converted to method' as follows:

"GET" to `GET

"HEAD" to `HEAD

"POST" to `POST

"PUT" to `PUT

"DELETE" to `DELETE

"CONNECT" to `CONNECT

"OPTIONS" to `OPTIONS

"TRACE" to `TRACE

Any other value m to `Method m

Note String comparison is case insensitive.

Wtr.method' "GET" = `GET;;
Wtr.method' "get" = `GET;;
Wtr.method' "method" = `Method "method"

Route and Router

val route : ?method':method' -> ('a, 'b) request_target -> 'a -> 'b route

route ~method' request_target handler is a route. The default value for ?method is `GET.

val routes : method' list -> ('a, 'b) request_target -> 'a -> 'b route list

routes methods request_target handler is a list of routes in which all have the same request_target and route handler value but each have one method' from methods. This is equivalent to calling route like so:

List.map (fun m -> route ~method:m request_target handler) [meth1; meth2; meth3]

val router : 'a route list -> 'a router

router routes is a router made up of given routes.

val router' : 'a route list list -> 'a router

router' routes_list = router (List.concat routes_list)

val match' : method' -> string -> 'a router -> 'a option

match' method' request_target router is Some a if method' and request_target together matches one of the routes defined in router. Otherwise it is None.

Pretty Printers

val pp_request_target : Stdlib.Format.formatter -> ('a, 'b) request_target -> unit

val pp_method : Stdlib.Format.formatter -> method' -> unit

val pp_route : Stdlib.Format.formatter -> 'b route -> unit

val pp : Stdlib.Format.formatter -> 'a router -> unit